minification of html, css, js and images for developer docs
also fixed some broken markup

+++ developer-ci.sh
@@ -38,6 +38,7 @@ if [ -n "$GO_SERVER_URL" ]; then
 fi
 
 export PATH=$HOME/.node/bin:$PATH
+npm prune
 npm install
 export PATH=$(npm bin):$PATH
 
@@ -45,20 +46,23 @@ export PATH=$(npm bin):$PATH
   cd $book
   gitbook install
   gitbook build .
+  grunt --build=developer
   rm -rf $HOME/.gocd-$book-docs
   mv _book $HOME/.gocd-$book-docs
 )
 
-git clean -dffx
-git fetch --all
-git branch -D gh-pages || true
-git checkout -b gh-pages origin/gh-pages
-git clean -dffx
+if [ -n "$PUSH_CHANGES" ]; then
+  git clean -dffx
+  git fetch --all
+  git branch -D gh-pages || true
+  git checkout -b gh-pages origin/gh-pages
+  git clean -dffx
 
-rm -rf $book
-mv $HOME/.gocd-$book-docs $book
+  rm -rf $book
+  mv $HOME/.gocd-$book-docs $book
 
-git add --all $book
-git commit -m "Updating site to latest commit ($git_short_sha)." --author "GoCD <go-cd-dev@googlegroups.com>"
+  git add --all $book
+  git commit -m "Updating site to latest commit ($git_short_sha)." --author "GoCD <go-cd-dev@googlegroups.com>"
 
-git push https://${GITHUB_USERNAME}:${GITHUB_PASSWORD}@github.com/gocd/documentation gh-pages:gh-pages
+  git push https://${GITHUB_USERNAME}:${GITHUB_PASSWORD}@github.com/gocd/documentation gh-pages:gh-pages
+fi
